tách static service

+++ micro-services/story-service/src/main/java/com/storyart/storyservice/dto/statistic/IRatingClassify.java
@@ -1,7 +1,7 @@
 package com.storyart.storyservice.dto.statistic;
 
 public interface IRatingClassify {
-int getStar();
+double getStar();
 int getCount();
 
 }
